add chainlink fetching, add dynamic monthly estima
