+++ packages/nextjs/package.json
@@ -35,7 +35,9 @@
     "@worldcoin/idkit": "^0.5.1",
     "daisyui": "^2.31.0",
     "date-fns": "^2.30.0",
+    "ethers": "^6.6.4",
     "framer-motion": "^10.13.0",
+    "fs": "^0.0.1-security",
     "mapbox-gl": "^2.15.0",
     "next": "^13.1.6",
     "nextjs-progressbar": "^0.0.16",
