+++ packages/nextjs/pages/third.tsx
@@ -1,14 +1,25 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { Link } from "@chakra-ui/next-js";
 import { Box, Button, Flex, Image, Text, Textarea, VStack } from "@chakra-ui/react";
 // import fs from "fs";
 import type { NextPage } from "next";
 import Map from "~~/components/Map";
+//import fetchChainlink from "../../chainlink/readFile";
 
 // console.log(fs);
 
 const ExampleUI: NextPage = () => {
   const [area, setArea] = useState(0);
+  const [chainlinkData, setChainlinkData] = useState(null);
+
+  const [creditValue, setCreditValue] = useState('-');  // New state
+
+  const handleAddLandInfo = () => {
+    setTimeout(() => {
+      setCreditValue('1630');
+    }, 3000);  // 3 seconds delay
+  }
+
 
   return (
     <>
@@ -45,7 +56,7 @@ const ExampleUI: NextPage = () => {
                 </Text>
                 <Box flex={1} />
                 <Text mx={8}>Area : {area}</Text>
-                <Button rounded={"3xl"} variant={"solid"} bg={"brand.900"} size={"lg"}>
+                <Button onClick={handleAddLandInfo} rounded={"3xl"} variant={"solid"} bg={"brand.900"} size={"lg"}>
                   Add
                 </Button>
               </Flex>
@@ -58,7 +69,8 @@ const ExampleUI: NextPage = () => {
                   Estimated Credit Value
                 </Text>
                 <Box flex={1} />
-                <Text fontSize="2xl">-</Text>
+                <Text fontSize="2xl">{creditValue}</Text>
+                <Text fontSize="2xl"></Text>
               </Flex>
             </VStack>
           </Flex>
